+++ src/record.rs
@@ -7,11 +7,11 @@ use std::{
 };
 
 use libc::{c_char, c_void};
-use slow5lib_sys::{slow5_file, slow5_rec_free, slow5_rec_t};
+use slow5lib_sys::{slow5_file, slow5_rec_free, slow5_rec_t, slow5_aux_get_enum};
 use thiserror::Error;
 
 use crate::{
-    auxiliary::{AuxField, AuxFieldSetExt},
+    auxiliary::{AuxField, AuxFieldSetExt, EnumField},
     error::Slow5Error,
     to_cstring, FileWriter,
 };
@@ -247,7 +247,6 @@ impl Record {
     }
 
     /// Get data for an auxiliary field of a record.
-    ///
     /// # Example
     /// ```
     /// # use anyhow::Result;
@@ -263,15 +262,37 @@ impl Record {
     /// # Ok(())
     /// # }
     /// ```
+    /// 
+    /// # Note
+    /// For fields of enum type, do not use use get_aux_field with T = u8, instead use [`get_aux_enum_field`].
     ///
     /// # Errors
     /// Returns an Err if auxiliary field wasn't set for that record.
+    /// 
+    /// [`get_aux_enum_field`]: crate::Record::get_aux_enum_field
     pub fn get_aux_field<B, T>(&self, name: B) -> Result<T, Slow5Error>
     where
         B: Into<Vec<u8>>,
         T: AuxField,
     {
         T::aux_get(self, name)
+    }
+
+    /// Get value for an enum field in a Record.
+    /// 
+    /// See [`EnumField`] to get more information about converting it into the enum label
+    pub fn get_aux_enum_field<B>(&self, name: B) -> Result<EnumField, Slow5Error>
+    where
+        B: Into<Vec<u8>>,
+    {
+        let mut err = 0;
+        let name = to_cstring(name)?;
+        let ef = unsafe { slow5_aux_get_enum(self.slow5_rec, name.as_ptr(), &mut err)};
+        if err < 0 {
+            Err(Slow5Error::Unknown)
+        } else {
+            Ok(EnumField(ef as usize))
+        }
     }
 }
 
@@ -541,11 +562,11 @@ impl RecPtr for Record {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::{auxiliary::FieldType, FileWriter};
+    use crate::{auxiliary::FieldType, FileWriter, FileReader};
+        use assert_fs::{fixture::PathChild, TempDir};
 
     #[test]
     fn test_aux() -> anyhow::Result<()> {
-        use assert_fs::{fixture::PathChild, TempDir};
         let tmp_dir = TempDir::new()?;
         let path = "new.slow5";
         let path = tmp_dir.child(path);
@@ -563,5 +584,13 @@ mod test {
             .build()?;
         rec.set_aux_field(&mut slow5, "median", 10.0f32)?;
         Ok(())
+    }
+
+    #[test]
+    fn test_get_aux_enum() {
+        let fp = "examples/example3.blow5";
+        let mut reader = FileReader::open(fp).expect("Unable to open example3.blow5");
+        let rec = reader.records().next().unwrap().unwrap();
+        let end_reason = rec.get_aux_enum_field("end_reason").expect("Unable to get end_reason enum field");
     }
 }